+++ cats/executor/structure/__init__.py
@@ -1,10 +1,6 @@
-class Structure:
-    def __init__(self,
-        service=None
-    ):
+class InfraStructure:
+    def __init__(self, service):
         self.service = service
-        self.bom_json_cid = self.service.bom_json_cid
-        # self.plant: Plant = plant
         # self.infraStructure: Terraform = Terraform(working_dir=cats.CWD)
 
     def destroy(self):
@@ -13,9 +9,14 @@ class Structure:
         print()
         print()
 
+    def plan(self):
+        print('Plan Structure!')
+        self.service.executeCMD(['terraform', 'plan'], cwd=self.service.CATS_HOME)
+        print()
+        print()
+
     def initialize(self):
         print('Initialize Structure!')
-        # self.service.executeCMD(['terraform', 'plan'])
         self.service.executeCMD(['terraform', 'init', '--upgrade'], cwd=self.service.CATS_HOME)
         print()
         print()
@@ -25,11 +26,3 @@ class Structure:
         self.service.executeCMD(['terraform', 'apply', '--auto-approve'], cwd=self.service.CATS_HOME)
         print()
         print()
-
-    def redeploy(self):
-        print()
-        print()
-        print('Deploy Structure!')
-        self.destroy()
-        self.initialize()
-        self.apply()
