--- cats/io/input/__init__.py
@@ -1,18 +1,0 @@
-from cats.io.input.function import InfraFunction, Processor
-
-class Function(InfraFunction):
-    def __init__(self, service):
-        self.service = service
-        self.infraFunc = InfraFunction.__init__(self, self.service)
-        self.orderCID: str = self.service.orderCID
-        self.processCID = self.service.integrated_subproc_cid
-        self.process: Processor = None
-        self.dataCID: str = None
-        self.invoiceCID: str = None
-
-    def execute(self):
-        self.process = self.infraFunc.configureProcess()
-        self.dataCID = self.process.execute(self.processCID)
-        self.invoiceCID = self.service.createInvoice(self.orderCID, self.dataCID, self.service.seedCID)
-        return self.invoiceCID
-
