--- cats/network/aws.py
@@ -1,2 +1,0 @@
-from cats import S3_CLIENT
-s3_client = S3_CLIENT
